rename useAuth hook and update imports
