+++ src/components/academy/CourseCard.tsx
@@ -12,7 +12,7 @@ import {
   BookmarkCheck,
   Play
 } from "lucide-react";
-import { useAuth } from "@/hooks/useAuth";
+import { useAuth } from "@/context/AuthContext";
 import { supabase } from "@/integrations/supabase/client";
 import { useToast } from "@/hooks/use-toast";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
