+++ src/hooks/useLegacyAuth.ts
